+++ proof_key.go
@@ -127,28 +127,3 @@ func ReadKeyProof(data []byte) (KeyProof, error) {
 	}
 	return nil, errors.New("unrecognized proof")
 }
-
-///////////////////////////////////////////////////////////////////////////////
-
-// InnerKeyProof represents a proof of existence of an inner node key.
-type InnerKeyProof struct {
-	*KeyExistsProof
-}
-
-// Verify verifies the proof is valid and returns an error if it isn't.
-func (proof *InnerKeyProof) Verify(hash []byte, value []byte, root []byte) error {
-	if !bytes.Equal(proof.RootHash, root) {
-		return errors.WithStack(ErrInvalidRoot)
-	}
-	if hash == nil || value != nil {
-		return errors.WithStack(ErrInvalidInputs)
-	}
-	return proof.PathToKey.verify(hash, root)
-}
-
-// ReadKeyInnerProof will deserialize a InnerKeyProof from bytes.
-func ReadInnerKeyProof(data []byte) (*InnerKeyProof, error) {
-	proof := new(InnerKeyProof)
-	err := wire.ReadBinaryBytes(data, &proof)
-	return proof, err
-}
